chatbot/requirements.txt
line-bot-sdk
flask
configparser
azure-ai-textanalytics